pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "asaca"
version = "0.1.0"
description = "Automatic Speech Analysis for Cognitive Assessment"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Xinbo Yang", email = "xyang2@tcd.ie"}]

dependencies = [
    "numpy>=1.24",
    "pandas>=1.5",
    "scipy>=1.10",
    "joblib>=1.3",
    "tqdm>=4.66",
    "sympy>=1.12",
    "torch>=2.0",
    "torchaudio>=2.0",
    "scikit-learn>=1.4",
    "transformers>=4.38",
    "datasets>=2.18",
    "evaluate>=0.4",
    "safetensors>=0.4",
    "soundfile>=0.12",
    "librosa>=0.10",
    "resampy>=0.4",
    "ctc-segmentation>=1.7",
    "ctc-decoder @ git+https://github.com/githubharald/CTCDecoder.git@6b5c3dd",
    "webrtcvad",
    "praat-parselmouth>=0.4",
    "pronouncing>=0.2",
    "pyctcdecode>=0.5",
    "jiwer>=3.0",
    "pyannote.audio>=3.1",
    "pyannote.core>=5.0",
    "matplotlib>=3.7",
    "pillow>=10.2",
    "shap>=0.44",
    "psutil>=5.9",
    "openpyxl>=3.1",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
# GPU-enabled extras
gpu = ["torch", "torchaudio", "pyannote.audio"]
# Development and testing
dev = ["ruff", "black", "pre-commit", "pytest", "pytest-cov", "mypy"]
# Documentation build
docs = ["mkdocs-material", "mkdocstrings[python]"]

[project.scripts]
asaca-cli = "asaca.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
[tool.ruff.lint]
select = []

[tool.mypy]
ignore_errors = true
